Split channel types
